+++ R/aaa.R
@@ -5,7 +5,7 @@
 #' @importFrom hydroGOF mNSE NSE pbias rmse nrmse rsr
 #' @importFrom Rcpp sourceCpp
 #' @importFrom spatstat.geom as.owin as.ppp nncross
-#' @importFrom stats median optimize setNames
+#' @importFrom stats median optim optimize setNames
 #' @importFrom yaml read_yaml write_yaml
 #' @importFrom utils modifyList packageVersion
 #' @useDynLib RPhosFate